# backend-golang/api/docs/swagger.yaml
definitions:
  dtos.MessageToProcessInDTO:
    properties:
      messageName:
        type: string
    type: object
  dtos.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  entities.NewOfferInDTO:
    properties:
      distributor_id:
        type: integer
      duration:
        type: number
      message:
        type: string
      shipping_cost:
        type: number
    type: object
  entities.NewRouteDetailInDTO:
    properties:
      coordinates:
        type: string
    type: object
  entities.PackageInDTO:
    properties:
      description:
        type: string
      dimensions:
        type: string
      image_url:
        type: string
      name:
        type: string
      value:
        type: number
    type: object
  entities.ShippingInDTO:
    properties:
      dest_addr:
        type: string
      details:
        type: string
      origin_addr:
        type: string
      package_id:
        type: integer
      recipient_id:
        type: integer
      sender_id:
        type: integer
    type: object
  entities.ShippingInPatchDTO:
    properties:
      selected_offer_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /distributor/offers:
    get:
      consumes:
      - application/json
      description: Get offers associated to a distributor from the database by passing
        an ID
      parameters:
      - description: Distributor ID
        in: query
        name: distributorId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get offers made by a distributor
  /distributor/offers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete offer from the database by passing an ID
      parameters:
      - description: Offer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Delete offer
  /distributor/requests:
    get:
      description: Get all requests which are stored in the database and expect offers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get all pending shipping requests
  /distributor/requests/{id}/offers:
    get:
      consumes:
      - application/json
      description: Get offers associated to a request from the database by passing
        an ID
      parameters:
      - description: Shipping request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get request offers
    post:
      consumes:
      - application/json
      description: Creates a new shipping request offer instance and stores it in
        the database
      parameters:
      - description: Fill the body to create a new shipping request offer
        in: body
        name: Offer
        required: true
        schema:
          $ref: '#/definitions/entities.NewOfferInDTO'
      - description: Shipping request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Creates a new offer for a shipping request
  /distributor/shippings/{id}:
    get:
      consumes:
      - application/json
      description: Get Shipping stored in the database by passing an ID
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get Shipping
    patch:
      consumes:
      - application/json
      description: Change shipping state by sending a message to a camunda process
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fill the body to change shipping state
        in: body
        name: Shipping
        required: true
        schema:
          $ref: '#/definitions/dtos.MessageToProcessInDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Update Shipping State
  /distributor/shippings/{id}/route:
    post:
      consumes:
      - application/json
      description: Add new coordinates to a shipping and stores it in the database
      parameters:
      - description: Fill the body to update coordinates
        in: body
        name: Offer
        required: true
        schema:
          $ref: '#/definitions/entities.NewRouteDetailInDTO'
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Create new coordinates
  /packages:
    post:
      consumes:
      - application/json
      description: Creates a new package in the database
      parameters:
      - description: Fill the body to create a new package
        in: body
        name: Package
        required: true
        schema:
          $ref: '#/definitions/entities.PackageInDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Creates a new package
  /packages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Package by ID which is stored in the database
      parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Delete Package
  /sender/shippings:
    get:
      description: Get all shippings stored in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get all shippings
    post:
      consumes:
      - application/json
      description: Creates a new shipping instance in camunda and in the database
      parameters:
      - description: Fill the body to create a new shipping
        in: body
        name: Shipping
        required: true
        schema:
          $ref: '#/definitions/entities.ShippingInDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Creates a new shipping
  /sender/shippings/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Shipping by ID which is stored in the database
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Delete Shipping
    get:
      consumes:
      - application/json
      description: Get Shipping stored in the database by passing an ID
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get Shipping
    patch:
      consumes:
      - application/json
      description: Change shipping state by sending a message to a camunda process
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fill the body to change shipping state
        in: body
        name: Shipping
        required: true
        schema:
          $ref: '#/definitions/dtos.MessageToProcessInDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Update Shipping State
  /sender/shippings/{id}/offers:
    get:
      consumes:
      - application/json
      description: Get offers associated to a shipping from the database by passing
        an ID
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get shipping offers
  /sender/shippings/{id}/offers/selected:
    get:
      consumes:
      - application/json
      description: Get a shipping selected route from the database by passing an ID
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get Shipping selected offer
    patch:
      consumes:
      - application/json
      description: Add selected offer and send a message to a camunda process
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      - description: Add selected offer ID
        in: body
        name: Shipping
        required: true
        schema:
          $ref: '#/definitions/entities.ShippingInPatchDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Update Shipping's selected offer
  /sender/shippings/{id}/route:
    get:
      consumes:
      - application/json
      description: Get a shipping route from the database by passing an ID
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get Shipping route
  /shippings/{id}/state:
    get:
      consumes:
      - application/json
      description: Get Shipping state making reference to Camunda instance
      parameters:
      - description: Shipping ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.Response'
      summary: Get Shipping state
swagger: "2.0"